change schedule to events
